Clean up Button component
