+++ Sources/Components/Button/ButtonSize.swift
@@ -2,7 +2,7 @@ import Foundation
 
 extension Warp {
     /// Warp button sizes.
-    public enum ButtonSize: Equatable, Comparable {
+    public enum ButtonSize: String, Equatable, CaseIterable {
         case big
         case small
     }
